+++ packages/eslint-plugin-security-practices/src/rules/no-patent-licenses.ts
@@ -20,7 +20,7 @@ module.exports = {
 
   create(context) {
     if (path.basename(context.getFilename()) !== 'package.json') {
-      return;
+      return {};
     }
 
     const cwd = context.getCwd();